--- src/main/java/org/kovan/intern/javafundamentals/control_flow_and_algorithmic_logic/BlueJPattern.java
@@ -1,12 +1,0 @@
-public class BlueJPattern{
-    public static void main(String[] args) {
-        String st = "ABCDE";
-
-        for(int i=0;i<st.length();i++){
-            for(int j=i;j<st.length();j++){
-                System.out.print(st.charAt(j)+" ");
-            }
-            System.out.println();
-        }
-    }
-}
