--- src/main/java/org/kovan/intern/javafundamentals/control_flow_and_algorithmic_logic/PerfectDiamond.java
@@ -1,25 +1,0 @@
-public class PerfectDiamond{
-	public static void main(String[] args) {
-       int m=7;
-       int n=m/2;
-       for(int i=0;i<=n;i++){
-           for(int j=0;j<n-i;j++){
-               System.out.print(" ");
-           }
-           for(int j=0;j<(2*i)+1;j++){
-               System.out.print("*");
-           }
-           System.out.println();
-       }
-      // System.out.println();
-       for(int i=n-1;i>=0;i--){
-           for(int j=0;j<n-i;j++){
-               System.out.print(" ");
-           }
-           for(int j=0;j<(2*i)+1;j++){
-               System.out.print("*");
-           }
-           System.out.println();
-       }
-    }
-}
