--- src/main/java/org/kovan/intern/javafundamentals/control_flow_and_algorithmic_logic/TernaryOperator.java
@@ -1,7 +1,0 @@
-public class TernaryOperator{
-	public static void main(String[] args) {
-	int a =10;
-	int b=5;
-	System.out.println(a>b?"A Greater":"B Greater");	
-	}
-}
