--- src/main/java/org/kovan/intern/javafundamentals/methods_and_recursion/StackOverFlowTest.java
@@ -1,8 +1,0 @@
-public class StackOverFlowTest{
-	public static void stackTest(){
-		stackTest();
-	}
-	public static void main(String[] args) {
-		stackTest();
-	}
-}
