--- src/main/java/org/kovan/intern/javafundamentals/strings_and_regular_expressions/RegexReplace.java
@@ -1,9 +1,0 @@
-package org.kovan.intern.javafundamentals.strings_and_regular_expressions;
-
-public class RegexReplace {
-    public static void main(String[] args) {
-        String input = "Kovan123Labs456";
-        String result = input.replaceAll("\\d+", "#");
-        System.out.println(result);
-    }
-}
