--- src/main/java/org/kovan/intern/javafundamentals/syntax_and_memory/OverflowTest.java
@@ -1,11 +1,0 @@
-package org.kovan.intern.javafundamentals.syntax_and_memory;
-
-public class OverflowTest{
-    public static void main(String[] args) {
-        byte bt = 126;
-        for(int i=0;i<5;i++){
-            System.out.println("value: " + bt);
-            bt++;
-        }
-    }
-}
