+++ src/main/java/org/kovan/intern/service/javafundamentals/methods_and_recursion/StackOverFlowTest.java
@@ -1,0 +1,10 @@
+package org.kovan.intern.service.javafundamentals.methods_and_recursion;
+
+public class StackOverFlowTest{
+	public static void stackTest(){
+		stackTest();
+	}
+	public static void main(String[] args) {
+		stackTest();
+	}
+}
